# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Voice Agent Bridge API
  version: 0.1.0
  description: |
    HTTP API for Voice Agent Bridge (VAB).

    **Security**
    - `UserBearer` — JWT from `/v1/auth/register` or `/v1/auth/login`
    - `AgentKey` — header `X-Agent-Key` (per-agent key from create/pairing claim)

    **Push semantics**
    - `POST /v1/sessions/{id}/progress` (**update_progress**) NEVER pushes
    - `POST /v1/sessions/{id}/events` (**report_event**) MAY push for `needs_user`,
      or `succeeded`/`failed` with `actions` / `force_push`
    - Agent decides `needs_user`; the bridge does not guess

    **TTL**
    - Session default max 24h (`86400`)
    - Destructive confirm window max 30m (`1800`)
servers:
  - url: http://127.0.0.1:8787
    description: Local MVP
...
paths:
...
  /v1/skills:
    get:
      tags: [skills]
      operationId: listSkills
      summary: List skills
      security:
        - UserBearer: []
        - AgentKey: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [skills]
                properties:
                  skills:
                    type: array
                    items:
                      $ref: "#/components/schemas/Skill"
    post:
      tags: [skills]
      operationId: upsertSkill
      summary: Create or upsert a skill
      security:
        - UserBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Skill"
      responses:
        "200":
          description: Upserted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
  /v1/sessions:
    post:
      tags: [sessions]
      operationId: createOrResumeSession
      summary: Create or resume a session (agent)
      description: Session is created when the agent calls this with `skill_id`.
      security:
        - AgentKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [skill_id]
              additionalProperties: false
              properties:
                skill_id: { type: string }
                session_id: { type: string }
                idempotency_key: { type: string }
                chat_id: { type: string }
                title: { type: string }
                facts:
                  type: object
                  additionalProperties: true
                metadata:
                  type: object
                  additionalProperties: true
      responses:
        "200":
          description: Created or resumed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "404":
          $ref: "#/components/responses/Error"
...
  /v1/phone/sessions:
    get:
      tags: [phone]
      operationId: listPhoneSessions
      summary: List non-closed sessions for the phone user
      security:
        - UserBearer: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [sessions]
                properties:
                  sessions:
                    type: array
                    items:
                      $ref: "#/components/schemas/Session"
  /v1/phone/sessions/{session_id}/history:
    get:
      tags: [phone]
      operationId: listSessionHistory
      summary: List the auditable lifecycle of one session
      security:
        - UserBearer: []
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 100 }
      responses:
        "200":
          description: Audit timeline
          content:
            application/json:
              schema:
                type: object
                required: [entries]
                properties:
                  entries:
                    type: array
                    items: { $ref: "#/components/schemas/HistoryEntry" }
  /v1/phone/devices:
    post:
      tags: [phone]
      operationId: registerPhoneDevice
      summary: Register a device for push (dev mode stores token)
      security:
        - UserBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [platform, push_token]
              additionalProperties: false
              properties:
                platform:
                  type: string
                  enum: [ios, ios_simulator, mac_dev]
                push_token: { type: string }
                locale:
                  type: string
                  default: zh-Hans
      responses:
        "200":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
  /v1/phone/sessions/{session_id}/reply:
    post:
      tags: [phone]
      operationId: replyPhoneSession
      summary: User selects a skill action from the phone
      description: |
        - low risk → action `queued`, session `queued`
        - destructive / confirm → action `pending_confirm`, session `awaiting_confirm`
          until `/confirm`
        - `confirm: false` records the phone cancellation as an agent-claimable
          queued action with `cancelled_by_user: true`.
      security:
        - UserBearer: []
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action_key]
              additionalProperties: false
              properties:
                action_key:
                  type: string
                  description: Skill action id (e.g. rollback, ack)
                utterance: { type: string }
      responses:
        "200":
          description: Reply accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhoneActionResponse"
  /v1/phone/sessions/{session_id}/confirm:
    post:
      tags: [phone]
      operationId: confirmPhoneSession
      summary: Second confirmation for destructive / confirm-required actions
      security:
        - UserBearer: []
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action_id, confirm]
              additionalProperties: false
              properties:
                action_id: { type: string }
                confirm: { type: boolean }
      responses:
        "200":
          description: Confirmation applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhoneActionResponse"
        "409":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    UserBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: User JWT from /v1/auth/register or /v1/auth/login
    AgentKey:
      type: apiKey
      in: header
      name: X-Agent-Key
      description: Per-agent API key from agent create or pairing claim
  parameters:
    SessionId:
      name: session_id
      in: path
      required: true
      schema: { type: string }
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    ProgressStatus:
      type: string
      enum: [started, running, blocked, succeeded, failed, cancelled]
      description: Progress update status — NEVER triggers push
...
    SessionState:
      type: string
      enum:
        - open
        - running
        - needs_user
        - awaiting_confirm
        - queued
        - claimed
        - closed
        - expired
    ActionRisk:
      type: string
      enum: [low, medium, high, destructive]
      description: destructive (or confirm:true) requires second confirm (≤30m)
    ActionStatus:
      type: string
      enum:
        - offered
        - pending_confirm
        - awaiting_confirm
        - queued
        - claimed
        - done
        - failed
        - cancelled
        - expired
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
        message: { type: string }
        details: { type: object, additionalProperties: true }
...
    HistoryEntry:
      type: object
      required: [audit_id, action, created_at, metadata]
      properties:
        audit_id: { type: string }
        action: { type: string }
        session_id: { type: [string, "null"] }
        agent_id: { type: [string, "null"] }
        metadata: { type: object, additionalProperties: true }
        created_at: { type: string, format: date-time }
    SkillActionDef:
      type: object
      required: [id, risk, confirm]
      properties:
        id: { type: string }
        risk:
          $ref: "#/components/schemas/ActionRisk"
        confirm: { type: boolean }
        title: { type: string }
    Skill:
      type: object
      required: [skill_id, template, facts_schema, actions, ttl]
      properties:
        skill_id: { type: string, examples: [deploy.result] }
        template:
          type: string
          examples: ["{{service}} 在 {{env}} 部署{{status}}"]
        facts_schema:
          type: array
          items: { type: string }
        actions:
          type: array
          items:
            $ref: "#/components/schemas/SkillActionDef"
        ttl:
          type: object
          required: [default_sec, destructive_sec]
          properties:
            default_sec:
              type: integer
              maximum: 86400
            destructive_sec:
              type: integer
              maximum: 1800
        version: { type: integer }
    Session:
      type: object
      required: [session_id, skill_id, state]
      properties:
        session_id: { type: string }
        user_id: { type: string }
        agent_id: { type: string }
        skill_id: { type: string }
        state:
          $ref: "#/components/schemas/SessionState"
        chat_id: { type: ["string", "null"] }
        title: { type: ["string", "null"] }
        progress_status:
          oneOf:
            - $ref: "#/components/schemas/ProgressStatus"
            - type: "null"
        progress_percent: { type: ["number", "null"], minimum: 0, maximum: 100 }
        progress_message: { type: ["string", "null"] }
        summary_text: { type: ["string", "null"] }
        voice_script:
          type: ["string", "null"]
          description: Summary truncated ~80 chars for voice
        available_actions:
          type: array
          items: { type: string }
        facts:
          type: object
          additionalProperties: true
        expires_at: { type: string }
        created_at: { type: string }
        updated_at: { type: string }
...
    Action:
      type: object
      required: [action_id, session_id, action_key, risk, status]
      properties:
        action_id: { type: string }
        session_id: { type: string }
        action_key: { type: string }
        title: { type: [string, "null"] }
        risk:
          $ref: "#/components/schemas/ActionRisk"
        confirm_required: { type: boolean }
        status:
          $ref: "#/components/schemas/ActionStatus"
        payload:
          type: object
          additionalProperties: true
        expires_at: { type: string }
        claimed_at: { type: ["string", "null"] }
        result:
          type: [object, "null"]
          additionalProperties: true
        cancelled_by_user: { type: boolean }
...
    Device:
      type: object
      required: [device_id, platform]
      properties:
        device_id: { type: string }
        platform: { type: string }
        push_token: { type: ["string", "null"] }
        locale: { type: string }
    PhoneActionResponse:
      type: object
      required: [session, action, needs_confirm]
      properties:
        session:
          $ref: "#/components/schemas/Session"
        action:
          $ref: "#/components/schemas/Action"
        needs_confirm: { type: boolean }
...